refactor(server): trim runner-profile cleanup theatre
Follow-up cleanup on c40612aa2 — strip change-detectors, paraphrasing
comments, dead branches, and the Map.copyOf ordering bug. Net −133
lines across 11 files; no behaviour change.

- Map.copyOf in MentorRunnerProfile.ENV doesn't preserve LinkedHashMap
  insertion order; switch to Collections.unmodifiableMap so the
  per-process env fragment stays deterministic build-to-build.
- Drop the unreachable null-key check in PiPlanSpec — Map.copyOf throws
  on null keys first, so the explicit branch was dead.
- Consolidate spec-validation tests into PiPlanSpecValidationTest;
  delete the duplicate SpecValidation nested class from
  PiRuntimeFactoryTest (it was testing the record, not the factory).
- Drop PR-narrating javadoc blocks ("This was previously a runner-
  script filename dispatch…", dead-flag graveyard, sealing rationale)
  and the `void ASPECT_INPUT_DIR; // intentionally referenced` hack.
- WorkspaceAbiSyncTest now pins values via regex (`=\\s*42\\b`) instead
  of JS identifier names, so renaming the runner const doesn't fail the
  test on a semantically-identical runner.
- Replace the magic-substring "jsonrpc" assertion in
  mentorProfileResolvesCorrectScript with the structural invariant —
  mentor and practice runner bytes differ.
- Drop the redundant plausibleButUndeclaredPathRejected and tautological
  nullKeyRejected tests (the former duplicated undeclaredPathRejected;
  the latter pinned Map.copyOf semantics, not PiPlanSpec validation).
