+++ server/application-server/src/test/java/de/tum/in/www1/hephaestus/agent/runtime/AgentRuntimeBoundaryTest.java
@@ -102,10 +102,6 @@ class AgentRuntimeBoundaryTest extends HephaestusArchitectureTest {
         @Test
         @DisplayName("PiRunnerProfile implementations reside in agent.practice or agent.mentor")
         void profilesInDomainPackages() {
-            // The profile interface is intentionally NOT sealed (a sealed interface with
-            // `permits PracticeRunnerProfile, MentorRunnerProfile` would force agent.runtime to
-            // import its domain implementations — exactly the dependency the boundary rules
-            // above forbid). Enforce the same constraint architecturally instead.
             ArchRule rule = classes()
                 .that()
                 .implement("de.tum.in.www1.hephaestus.agent.runtime.PiRunnerProfile")
